+++ lib/src/async_task_generic.dart
@@ -2,7 +2,7 @@ import 'async_task_base.dart';
 
 int getAsyncExecutorMaximumParallelism() => 1;
 
-AsyncExecutorThread? createMultiThreadAsyncExecutorThread(
+AsyncExecutorThread? createMultiThreadAsyncExecutorThread(String executorName,
     AsyncTaskLoggerCaller logger, bool sequential, int parallelism,
     [AsyncTaskRegister? taskRegister]) {
   return null;